app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
asyncpg
python-dotenv
psycopg2-binary


